# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Araponga API",
    "version": "v1",
    "description": "Araponga é uma plataforma comunitária orientada a território. Esta API expõe recursos de território, feed comunitário, entidades do mapa e saúde do território. O objetivo do MVP é viabilizar cadastro, visualização e curadoria comunitária com segurança e governança mínima.",
    "contact": {
      "name": "Araponga (maintainers)"
    },
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/api/v1/cart": {
      "get": {
        "tags": [
          "Cart"
        ],
        "summary": "Retorna o carrinho do usuário no território.",
        "parameters": [
          {
            "name": "territoryId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/cart/items": {
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "Adiciona um item ao carrinho.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartItemResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/cart/items/{id}": {
      "patch": {
        "tags": [
          "Cart"
        ],
        "summary": "Atualiza um item do carrinho.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartItemResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "summary": "Remove um item do carrinho.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/cart/checkout": {
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "Finaliza o checkout do carrinho.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartCheckoutResponse"
                },
                "examples": {
                  "checkout": {
                    "summary": "Checkout com checkouts e inquiries",
                    "value": {
                      "checkouts": [
                        {
                          "id": "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb",
                          "territoryId": "22222222-2222-2222-2222-222222222222",
                          "buyerUserId": "cccccccc-cccc-cccc-cccc-cccccccccccc",
                          "storeId": "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
                          "status": "CREATED",
                          "currency": "BRL",
                          "itemsSubtotalAmount": 65.0,
                          "platformFeeAmount": 3.25,
                          "totalAmount": 68.25,
                          "createdAtUtc": "2024-11-20T12:00:00Z",
                          "updatedAtUtc": "2024-11-20T12:00:00Z",
                          "items": [
                            {
                              "id": "dddddddd-dddd-dddd-dddd-dddddddddddd",
                              "listingId": "eeeeeeee-eeee-eeee-eeee-eeeeeeeeeeee",
                              "listingType": "PRODUCT",
                              "titleSnapshot": "Mel artesanal 500g",
                              "quantity": 2,
                              "unitPrice": 32.5,
                              "lineSubtotal": 65.0,
                              "platformFeeLine": 3.25,
                              "lineTotal": 68.25,
                              "createdAtUtc": "2024-11-20T12:00:00Z"
                            }
                          ]
                        }
                      ],
                      "inquiries": [
                        {
                          "id": "ffffffff-ffff-ffff-ffff-ffffffffffff",
                          "storeId": "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
                          "listingId": "11111111-1111-1111-1111-111111111111",
                          "batchId": "22222222-2222-2222-2222-222222222222"
                        }
                      ],
                      "summaries": [
                        {
                          "storeId": "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
                          "checkoutItemCount": 1,
                          "inquiryCount": 1
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StoreContactPayload": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "nullable": true
          },
          "whatsapp": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "instagram": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "preferredContactMethod": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "StoreResponse": {
        "required": [
          "id",
          "territoryId",
          "ownerUserId",
          "displayName",
          "status",
          "paymentsEnabled",
          "contactVisibility",
          "createdAtUtc",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "paymentsEnabled": {
            "type": "boolean"
          },
          "contactVisibility": {
            "type": "string"
          },
          "contact": {
            "$ref": "#/components/schemas/StoreContactPayload"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ListingResponse": {
        "required": [
          "id",
          "territoryId",
          "storeId",
          "type",
          "title",
          "pricingType",
          "status",
          "createdAtUtc",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "pricingType": {
            "type": "string"
          },
          "priceAmount": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AddCartItemRequest": {
        "required": [
          "territoryId",
          "listingId",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "listingId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateCartItemRequest": {
        "required": [
          "quantity"
        ],
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CartItemResponse": {
        "required": [
          "id",
          "listingId",
          "quantity",
          "createdAtUtc",
          "updatedAtUtc",
          "listing",
          "store",
          "isPurchasable"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "listingId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "listing": {
            "$ref": "#/components/schemas/ListingResponse"
          },
          "store": {
            "$ref": "#/components/schemas/StoreResponse"
          },
          "isPurchasable": {
            "type": "boolean"
          }
        }
      },
      "CartResponse": {
        "required": [
          "cartId",
          "territoryId",
          "userId",
          "createdAtUtc",
          "updatedAtUtc",
          "items"
        ],
        "type": "object",
        "properties": {
          "cartId": {
            "type": "string",
            "format": "uuid"
          },
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemResponse"
            }
          }
        }
      },
      "CartCheckoutRequest": {
        "required": [
          "territoryId"
        ],
        "type": "object",
        "properties": {
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CheckoutItemResponse": {
        "required": [
          "id",
          "listingId",
          "listingType",
          "titleSnapshot",
          "quantity",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "listingId": {
            "type": "string",
            "format": "uuid"
          },
          "listingType": {
            "type": "string"
          },
          "titleSnapshot": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "lineSubtotal": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "platformFeeLine": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "lineTotal": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CheckoutResponse": {
        "required": [
          "id",
          "territoryId",
          "buyerUserId",
          "storeId",
          "status",
          "currency",
          "createdAtUtc",
          "updatedAtUtc",
          "items"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "buyerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "itemsSubtotalAmount": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "platformFeeAmount": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "totalAmount": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckoutItemResponse"
            }
          }
        }
      },
      "InquiryBundleResponse": {
        "required": [
          "id",
          "storeId",
          "listingId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "listingId": {
            "type": "string",
            "format": "uuid"
          },
          "batchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "CheckoutSummaryResponse": {
        "required": [
          "storeId",
          "checkoutItemCount",
          "inquiryCount"
        ],
        "type": "object",
        "properties": {
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "checkoutItemCount": {
            "type": "integer",
            "format": "int32"
          },
          "inquiryCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CartCheckoutResponse": {
        "required": [
          "checkouts",
          "inquiries",
          "summaries"
        ],
        "type": "object",
        "properties": {
          "checkouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckoutResponse"
            }
          },
          "inquiries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InquiryBundleResponse"
            }
          },
          "summaries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckoutSummaryResponse"
            }
          }
        }
      },
...
    }
  }
}